+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Racetrack client returns non-zero exit code in case of error.
+  [issue #224](https://github.com/TheRacetrack/racetrack/issues/224)
 
 ## [2.12.0] - 2023-03-29
 ### Added